Refactor logging and module imports for consistency.
- Renamed `imdb.py` to `imdb_lookup.py` for better module naming consistency
- Made some import statement changes in `imdb_lookup.py`, `log.py`, and `string_parser.py`
- Refactored `AdvTitleSearch` and `FindImdbID` classes in `imdb_lookup.py` without changing their functionalities
- Added new parameters and functions to `log.py` for thread safety and to support multi-language logging
- Added `LanguageData` to `CreateProviderUrls` constructor parameters and updated `imdb_lookup` function call in `CreateProviderUrls` in `string_parser.py`
